Fix issue where json wasn't resolving
Adding the github pages configuration made the app look in a different place for the public folder (/climate-risk-map), so I couldn't load json by relative path.

See https://create-react-app.dev/docs/using-the-public-folder for more info using the public folder

Note I'm not placing this with the javascript and using import because I couldn't figure out how to do that and give it the correct type.