Show grey color for missing data
Previously, we were coloring unknown values the same as 0 values. The newer version of d3-scale forwards unknown values through the interpolator, so we can assign a special color to them. For now I'm choosing a medium-grey color. I want it to be low-contrast, yet distinguishable from 0 values (which are often almost white).