Temporarily move climate moisture index to normalized data
Until we have real normalized data for climate stuff, we can use this as a hacky stand-in. Years and datasets aren't working for normalized data so I removed them from climate moisture index. The issue with them is that when I change them in the multidataselector, the home.tsx doesn't seem to update the state, or propagate the state change at least back to the multidataselector