Add default dates and dataset for visualizations
This way, certain maps can start out showing an arbitrary year (not like, always the most recent year), or an abitrary dataset, not just the first one in the dataset list.

If defaults aren't set, that's okay too! We can just choose the most recent, or only, date and dataset.