Run editor endpoints on a different port
For security, we don't want to expose these endpoints right now.
Since we're only forwarding to the port the app is running on,
it should be fine to run the editor in the container under a
different port number.

Need to add EDITOR_PORT to the .env file or as an environment variable