Remove threshold color palettes for now
See https://github.com/cypressf/mit-climate-data-viz/issues/217#issuecomment-1035519730

This is a quick fix for a color selection bug, where selecting the scale type would change the allowable color palettes and domains, which is annoying because we'd have to specify which domains and scale types are compatible with which color palettes.

All interpolators work the same as each other, so we can just use them as-is.

The only threshold scale-type in the database right now is on a bubble map, so it's not even being used, and is safe to remove.

When we bring it back, look at https://github.com/cypressf/mit-climate-data-viz/issues/217#issuecomment-1035519730 and https://github.com/cypressf/mit-climate-data-viz/issues/217#issuecomment-1035252394 and https://observablehq.com/@d3/color-schemes for some of the limitations on the different palettes.