Use css modules instead of global css
The editor css, uploader css, and css for the viewer were conflicting without my knowledge. If you import css into one, it imports it into all of them because they are all imported together in `App.tsx`.

It's worth investigating if we can literally split these into separate pages so we don't import all the dependencies of Home when we're just trying to view Editor.