Automatically visualize the true domain if not set
The domain can be manually adjusted, but we should pick a sane default if it's not set.

TODO: performance optimization. is this slowing down the website? should we sort each dataset on the server-side and return the domain with the data? should we get the domain only once when we first get the data, then chache it on the client side? should we calculate the domain when the data is first uploaded to the server, and save it?