Allow linking to a specific tab
Previously, the website could only be opened to the "Combined Data" tab. If you changed to another tab and tried to bookmark it or send someone the link, it would just load the combined data tab again.

Now you can link to, bookmark, and share a specific tab. I'm still storing the navigation state in redux to keep state in one location, but if the url contains a valid tab id, it will update the redux state.