Show uncategorized map visualizations
I also want people to be able to mess around with map visualizations without them appearing on the website. This allows "draft" map visualizations that are not visible to people viewing the website.

TODO:

- "add" button to create new uncategorized map visualizations
- ability to add a map visualization to a tab

I created an ephemeral "-1" data category that doesn't appear in the database, but the server will report all uncategorized map visualizations under this category. The client can specify whether they want such drafts or not when requesting the categories and the map visualizations.