Add dataset editor
Previously we couldn't change the dataset for a map visualization. Now we can.

It is important to clear the default data source and default start and end date for a map visualization when the dataset changes, becuase the new dataset probably doesn't have those dates or sources available. I should do this on the backend too, but I'll implement it when I change the `PATCH` api to only take a diff, not an entire map visualization object. That way I can recognize which properties are being changed, vs which are remaining constant.

There's a bug if you change the dataset of the default map visualization in any tab, because the default dataset and start and end date are hardcoded in the frontend. These defaults should be loaded from the backend. In addition, the default selection for each tab should be loaded from the backend.

I made `name` and `dataset_name` separate properties on `MapVisualization` because it's important to be able to edit the dataset and the name separately. For the convenience of the viewer, I added a `displayName` property.