Do not show map visualization if it's not in tab
When switching tabs, we would show the map visualization that was selected from the previous tab before the new tab loaded anything. This was bad because if the new tab had no map visualizations, or a bad default map visualization, it would allow someone to edit a map visualization in a different tab.

For example. if you opened risk metrics, then opened "uncategorized" you would still see the map you had selected in risk metrics, because "uncategorized" had no map selected. This allowed someone to edit a risk metrics map when viewing the uncategorized tab.

The map visualization request `isUninitialized` when the skip token gets passed in.