Refactor: Dynamically load tabs
Previously I was hardcoding tabs because it was easier to manager the redux state that way. Now that tabs can dynamically show or hide themselves depending on the world/usa filter, it's better to dynamically load the ones I need. This also allows us to customize tabs in the editor in the future.

This also introduces bubble-map colors into the database so we can change those in the editor as well. They were previously hardcoded on the frontend.

I still haven't checked the react render flow, so it's possible this introduces a performance regression wherein the map is rendered multiple times.

TODO: check performance