Combine all data into one table
Since all data is associated with a dataset, the id it has doesn't matter (county FIPS code, ISO country code, etc) as long as the dataset contains the metadata about the id format. Now that the dataset table has a geography_type column, it does.

This reduces duplicate code paths for things that handle data from the usa database vs the world database.